requirements.txt
langchain==0.1.11
langchain-openai==0.0.8
langchain-community==0.0.31
langchain-pinecone==0.0.3
openai==1.13.3
python-dotenv==1.0.1
pydantic==2.6.3
fastapi==0.110.0
uvicorn==0.27.1
requests==2.31.0
python-multipart==0.0.9
# pytest==8.1.1
minio==7.2.5
pyyaml==6.0.1
pinecone-client==3.2.2

-e ./

ipykernel==6.29.3
ipywidgets==8.1.2